requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pyTelegramBotAPI==4.21.0
requests==2.32.3
pydantic==2.8.2
APScheduler==3.10.4
SQLAlchemy==2.0.32
psycopg2-binary==2.9.9
python-dotenv==1.0.1
openai==1.43.0
